training/requirements.txt
scapy==2.5.0
tensorflow-cpu==2.16.1
prometheus_client==0.22.1
psutil==5.9.8
